+++ src/Document.php
@@ -11,15 +11,17 @@ use MongoDB\BSON\Unserializable;
 
 class Document implements Iterator, ArrayAccess, JsonSerializable, Serializable, Unserializable
 {
-    use Concerns\HasId;
     use Concerns\HasCollection;
+    use Concerns\HasId;
     use Concerns\HasRelationships;
 
-    use Document\MongoAccess;
     use Document\ArrayAccess;
     use Document\ArrayIterator;
-    use Document\Accessors\CamelCaseAccessor;
+    use Document\CamelCaseAccessor;
+    use Document\MongoAccess;
+
     use Document\Serializers\JsonSerializer;
+    use Document\Serializers\NativeSerializer;
 
     protected $withId = false;
     protected $document = [];