+++ src/Document/Serializers/NativeSerializer.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Xenus\Document\Serializers;
+
+trait NativeSerializer
+{
+    /**
+     * Specify which properties must be serialized
+     *
+     * @return array
+     */
+    public function __sleep()
+    {
+        return ['document'];
+    }
+}